Rename out of index.js, use component names instead
