+++ assets/js/components/Loading/Loading.js
