+++ assets/js/components/OnboardingSteps/Interests.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { FormattedMessage, injectIntl } from 'react-intl'
 import { connect } from 'react-redux'
 import { fetchAllInterests, fetchAllRegions } from '../../user/redux'
-import HelSelect from '../HelSelect'
+import HelSelect from '../HelSelect/HelSelect'
 
 import InterestsList from '../Interests/InterestsList'
