+++ assets/js/components/OnboardingSteps/ProfileImage.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import { Col, Row, Form, Button } from 'reactstrap'
 import { FormattedMessage, injectIntl } from 'react-intl'
-import HelTextInput from '../HelTextInput'
-import ImgDropAndCrop from '../ImgDropAndCrop'
+import HelTextInput from '../HelTextInput/HelTextInput'
+import ImgDropAndCrop from '../ImgDropAndCrop/ImgDropAndCrop'
 
 class ProfileImage extends Component {
   componentDidUpdate(prevProps) {
