+++ assets/js/components/OnboardingSteps/Settings.js
@@ -3,7 +3,7 @@ import { injectIntl } from 'react-intl'
 import PropTypes from 'prop-types'
 import { Form, FormGroup, Input, Label, Col, Row } from 'reactstrap'
 
-import If from '../If'
+import If from '../If/If'
 
 const toBool = str => str === 'true'
 
@@ -61,76 +61,82 @@ const Settings = ({
             </Col>
           </Row>
         </FormGroup>
-        <If condition={enableNotifications}>
-          <FormGroup
-            onChange={messagesHandler}
-            className="oma-settings__messages"
-          >
-            <h4>{intl.formatMessage({ id: 'onboarding.messages.heading' })}</h4>
-            <p>{intl.formatMessage({ id: 'onboarding.messages.text' })}</p>
-            <div className="check-group">
-              <Label check>
-                <Input
-                  type="checkbox"
-                  value="email"
-                  defaultChecked={enabledMessages.includes('email')}
-                />{' '}
-                {intl.formatMessage({ id: 'onboarding.label.email' })}
-              </Label>
-              <Label check>
-                <Input
-                  type="checkbox"
-                  value="sms"
-                  defaultChecked={enabledMessages.includes('sms')}
-                />{' '}
-                {intl.formatMessage({ id: 'onboarding.label.sms' })}
-              </Label>
-              <Label check>
-                <Input
-                  type="checkbox"
-                  value="app"
-                  defaultChecked={enabledMessages.includes('app')}
-                />{' '}
-                {intl.formatMessage({ id: 'onboarding.label.app' })}
-              </Label>
-            </div>
-          </FormGroup>
-          <FormGroup
-            onChange={notificationsHandler}
-            className="oma-settings__notifications"
-          >
-            <h4>
-              {intl.formatMessage({ id: 'onboarding.notifications.heading' })}
-            </h4>
-            <p>{intl.formatMessage({ id: 'onboarding.notifications.text' })}</p>
-            <div className="check-group">
-              <Label check>
-                <Input
-                  type="checkbox"
-                  value="event"
-                  defaultChecked={enabledNotifications.includes('event')}
-                />{' '}
-                {intl.formatMessage({ id: 'onboarding.label.notifyEvent' })}
-              </Label>
-              <Label check>
-                <Input
-                  type="checkbox"
-                  value="area"
-                  defaultChecked={enabledNotifications.includes('area')}
-                />{' '}
-                {intl.formatMessage({ id: 'onboarding.label.notifyArea' })}
-              </Label>
-              <Label check>
-                <Input
-                  type="checkbox"
-                  value="subject"
-                  defaultChecked={enabledNotifications.includes('subject')}
-                />{' '}
-                {intl.formatMessage({ id: 'onboarding.label.notifySubject' })}
-              </Label>
-            </div>
-          </FormGroup>
-        </If>
+        {enabledNotifications && (
+          <div>
+            <FormGroup
+              onChange={messagesHandler}
+              className="oma-settings__messages"
+            >
+              <h4>
+                {intl.formatMessage({ id: 'onboarding.messages.heading' })}
+              </h4>
+              <p>{intl.formatMessage({ id: 'onboarding.messages.text' })}</p>
+              <div className="check-group">
+                <Label check>
+                  <Input
+                    type="checkbox"
+                    value="email"
+                    defaultChecked={enabledMessages.includes('email')}
+                  />{' '}
+                  {intl.formatMessage({ id: 'onboarding.label.email' })}
+                </Label>
+                <Label check>
+                  <Input
+                    type="checkbox"
+                    value="sms"
+                    defaultChecked={enabledMessages.includes('sms')}
+                  />{' '}
+                  {intl.formatMessage({ id: 'onboarding.label.sms' })}
+                </Label>
+                <Label check>
+                  <Input
+                    type="checkbox"
+                    value="app"
+                    defaultChecked={enabledMessages.includes('app')}
+                  />{' '}
+                  {intl.formatMessage({ id: 'onboarding.label.app' })}
+                </Label>
+              </div>
+            </FormGroup>
+            <FormGroup
+              onChange={notificationsHandler}
+              className="oma-settings__notifications"
+            >
+              <h4>
+                {intl.formatMessage({ id: 'onboarding.notifications.heading' })}
+              </h4>
+              <p>
+                {intl.formatMessage({ id: 'onboarding.notifications.text' })}
+              </p>
+              <div className="check-group">
+                <Label check>
+                  <Input
+                    type="checkbox"
+                    value="event"
+                    defaultChecked={enabledNotifications.includes('event')}
+                  />{' '}
+                  {intl.formatMessage({ id: 'onboarding.label.notifyEvent' })}
+                </Label>
+                <Label check>
+                  <Input
+                    type="checkbox"
+                    value="area"
+                    defaultChecked={enabledNotifications.includes('area')}
+                  />{' '}
+                  {intl.formatMessage({ id: 'onboarding.label.notifyArea' })}
+                </Label>
+                <Label check>
+                  <Input
+                    type="checkbox"
+                    value="subject"
+                    defaultChecked={enabledNotifications.includes('subject')}
+                  />{' '}
+                  {intl.formatMessage({ id: 'onboarding.label.notifySubject' })}
+                </Label>
+              </div>
+            </FormGroup>
+          </div>
+        )}
       </Form>
     </div>
   )
