+++ assets/js/components/ServiceConsent/ServiceConsent.js
