+++ assets/js/components/Tab/index.js
@@ -1,6 +1,6 @@
-import Interest from './Interest'
-import Profile from './Profile'
-import Service from './Service'
-import History from './History'
+import Interest from './Interest/Interest'
+import Profile from './Profile/Profile'
+import Service from './Service/Service'
+import History from './History/History'
 
 export { Interest, Profile, Service, History }
